// projectmate/redux-store/actions/action-types.tsx
enum actionTypes{
    //for user Auth
    REQUESTEDSETUSER = "Request for set user for first time",
    REQUESTEDSIGNUP = "Request for signup",
    REQUESTEDSIGNIN = "Request for signin",
    REQUESTEDSIGNOUT = "Request for signout",

    SETUSER = "set user for first time",
    USERSIGNEDUP = "user signedup",
    USERSIGNEDIN = "user signedin",
    USERSIGNEDOUT = "user signedout"


}


export default actionTypes;